Rename ProfileViewWhenReady to AppViewRouter and move to its own file
